AG - Code refactored, move searchCase to utils and renamed step defs to follow naming conventions. Enum moved into class where it is used.
