--- src/test/java/uk/gov/hmcts/darts/automation/cucumber/steps/SearchResultsEnum.java
@@ -1,10 +1,0 @@
-package uk.gov.hmcts.darts.automation.cucumber.steps;
-
-public enum SearchResultsEnum {
-    CASEID,
-    COURTHOUSE,
-    COURTROOM,
-    JUDGES,
-    DEFENDANTS,
-    INVALID
-}
