+++ src/test/java/uk/gov/hmcts/darts/automation/cucumber/steps/StepDef_base.java
@@ -21,7 +21,6 @@ public class StepDef_base {
 	
 	final WebDriver webDriver;
 	NavigationShared NAV;
-	searchCase SRCH;
 	
 	public StepDef_base(
 			SharedDriver webDriver, 
@@ -29,14 +28,6 @@ public class StepDef_base {
 			) {
 		this.webDriver = webDriver;
 		this.NAV = NAV;
-	}
-
-	public StepDef_base(
-			SharedDriver webDriver,
-			searchCase SRCH
-	) {
-		this.webDriver = webDriver;
-		this.SRCH = SRCH;
 	}
 	
 }
